+++ types/TFoldersConfig.d.ts
@@ -2,16 +2,19 @@
  * @param config path to folder with config files
  * @param models path to folder with moidels files
  * @param controllers path to folder with controllers files
- * @param views path to folder with view files
+ * @param migrations path to folder with migrations files
  * @param locales path to folder with locales files
- * @param emails path to folder with emails files
+ * @param migrations path to folder with migrations files
+ * @param [emails] path to folder with emails files. Optional
  */
 type FolderConfig = {
   config: string;
   models: string;
   controllers: string;
-  views: string;
-  emails: string;
+  commands: string;
+  locales: string;
+  migrations: string;
+  emails?: string;
 };
 
 export default FolderConfig;