+++ services/messaging/index.js
@@ -1,3 +1,0 @@
-import email from './email/index.js';
-
-export { email };