+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### NEXT
 
+[BREAKING] move email module to separate package @adaptivestone/framework-module-email. Please use it if you want to send emails
 [NEW] app now contains 'frameworkFolder' folder the framework located. Mostly for modules usage
 [BREAKING] remove VIEWS folders at all. Should not afffect any user as this was not used internally
 [UPDATE] update typing